Use a better font for annotations
The new font is copied from a PDF I generated from Google Docs. It is
added to the resources of each page before the text operations are
added.

Unfortunately since this new font is a CID font the text is specified in
CID indices instead of ASCII characters, and I haven't worked out the
conversion between these so it's hardcoded in a dictionary I generated
by looking at the internals of font_donor.pdf. TODO: fix hack :D