Use "absolute" method for matching op lengths
The theory in this method is that differences in the lengths of
operations are due to quantisation of the endpoints, and thus introduce
an absolute length error on the order of the size of the quantisation
steps. The threshold of 0.2 loosely comes from:

quantisation error ~= 0.05
one error at each of two endpoints in reference operations
one error at each of two endpoints in document operations
=> maximum error of 0.2 at document scale

but in reality I just increased it until it worked, then came up with
the explanation.

This seems to fix the unreliable matching of rotated twos - it remains
to be seen if it brings in any other problems.